Testing CSS importing on a per module basis
This test is enlightening, but at the moment a failure of sorts. I tried side effect importing a CSS file into each JS module; per module, that appears to work. However, the CSS file linked to the module that is last imported inside index.js is applied to the entire page from then on, the buttons don't change which CSS file is used.

This issue could be easily resolved in this case by applying the style from the CSS file inside the createContent() function, but for more complex files that would be unreasonable.

Odin mentioned something about scoping CSS files to a certain set of files, and I assume this is a situation where that comes into play. However I'm uncertain if that would actually change the issue here since I need them to apply to index.js.

The key issue is to have the CSS from a module actually used when that module's function is imported, but I'd like to avoid doing that inside createContent().

Also as a last note: This will probably be the last major commit to this project, the process of getting webpack working is the point of this project and from I can tell I have succeeded the basic criteria. The restaurant page itself would just be a process of typing stuff out, and I feel comfortable with DOM manipulation so I don't feel the need to practice it more when I could move onto other content. I would like to figure out the CSS stuff though.

/src
- Create /pageSpecificFiles

/pageSpecificFiles
- Create home.css, about.css, and menu.css

home.css, about.css, and menu.css
- All just style the p with a different color

home.js, about.js, menu.js
- Each side effect import their respective CSS file
 - I wouldn't be surprised if the side effect import is a key part of the reason I am having issues with the last imported module only ever being applied